Support either long-names or short-names for properties
Prior to this change Flo only recognized use of the short names
and any use of long names would not get reflected in the
properties view.

With this change the DSL text can include either short or
long names and the properties view will correctly pickup any
specified values. When the properties view is closed the
DSL text will be updated to reflect new values and should use
the same form of name (long or short) that was originally
used. If the properties view is used to set a property
that previously had no value the DSL text will use the
short name.