Temporarily deactivate eager stream name validation in create tab
When building streams it is nice to eagerly validate stream
names and let the user know if they are specifying something
invalid as early as possible.  Unfortunately the current
endpoint logs an exception on the server when a check is made
for a name that does not currently exist. Until we can change
that to avoid the exception for our scenario, this commit
turns off validation.