Validation for the bulk task definition dialog
The parser has moved from stream services to shared
services because it is now also being used in task
definition parsing.

Validation uses a temporary cache to ensure we do
not ping the server about every use of the same app
when creating multiple definitions. Cache expires
at the end of validation.

Validation is a two step process:
- syntactic. Check the format of the text
- semantic. Check the named task application exists
  and that any options specified for it are supported by
  it.