Task defs list up to date after task create
resolves #476

This was a sporadic issue that only occured when using the H2 DB.
And could not be reproduced using the server provided by angular.
The source problem was that the redirect only navigate was outside
the Observable and it would fire before the observable was complete.

To reproduce the problem.
. Build UI using `mvn clean install`
. Update SCDF pom to use BUILD-SNAPSHOT of UI
. Build SCDF
. Start SCDF using the default H2 database
. Add task definitions.