Turn off blur action for dsl text field and update flo
Just turning off the blur action would be bad but the update
to flo ensures we pickup a flo change where graph-to-text
syncing is turned on once you start interacting with the graph.

In this way if you simply switch to a different browser tab and
back, the DSL text field will still retain focus and you won't
lose some text you have in progress that isn't yet complete.

Fixes #423