Ensure recognize tap streams when computing text from graph
Without this change if the first stream in a graph is a tap stream
the graph to text algorithm will make a mistake and not treat it
as a tap stream because Array.find returns 0 in this case (because
it is stream 0 that is a tap stream). 0 is treated as false. It
is better to check the Array.find call returned a result (i.e.
result is not undefined)

Fixes #603