Add client-side reverse proxy feature
When the UI is accessed over reverse-proxy the url looks like
`https://reverse-proxy-url:443/scdf/` with the actual backend
url like `http://somedomain.de:8080/scdf/`.

If you access `_links` the backend is rendering them with the
backend url rather than the reverse proxy url.

This commit exchanges the protocol / host / port for the
rendered links. For example, `http://somedomain.de:8080/scdf/`
is replaced with `https://reverse-proxy-url:443/scdf/`.

User's have to opt-in to this behavior (checkbox in settings page).
If the server does not run behind a reverse proxy then nothing
is changed.

Resolves #1994